requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlmodel>=0.0.16
pydantic>=2.5.0
pydantic-settings>=2.1.0
passlib>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
pyjwt>=2.8.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-multipart>=0.0.6
openai>=1.12.0